\name{simGO-package}
\alias{simGO-package}
\alias{simGO}
\docType{package}
\title{Simulation Framework for Genotypes, Omics, and Phenotypes}
\description{
Tools for simulating relationships among genotypes, omics, and phenotypes.
}
\details{
The \code{simGO} package supports genotype and phenotype simulation as well as
multi-ancestry polygenic risk score benchmark design.
}
\author{
Quanshun Mei
}
\keyword{package}
